New Version OOD
New version which fully utilises python objects, allowing for more efficient code and opening the way for more complex system design. The new structure follows a model of delegating specific tasks to objects which are built with the methods required to properly process and store the data. System state savings and historical tracking have yet to be implemented, but will soon be.